refactor(job_service, web_service): streamline job retrieval and processing logic
