+++ src/service/job_service.py
@@ -34,19 +34,18 @@ class JobService:
         jobs = []
         filtered_jobs: list[SearchResult] = []
 
-        jobs = self.web_service.get_job_tags(search_params, pages)
+        jobs = self.web_service.get_jobs(search_params,pages)
 
         for job in jobs:
-            clean_job = self.process_job(job)
             if not self.extraction_service.keywords:
                 if search_params.filter_locations:
-                    if any(location in clean_job.job_details.location.lower() for location in search_params.filter_locations):
-                        filtered_jobs.append(SearchResult(clean_job, []))
+                    if any(location in job.job_details.location.lower() for location in search_params.filter_locations):
+                        filtered_jobs.append(SearchResult(job, []))
                 else:
-                    filtered_jobs.append(SearchResult(clean_job, []))
-            elif self.extraction_service.detect_keywords(clean_job.description):
-                keywords = self.extraction_service.extract_keywords(clean_job.description)
-                filtered_jobs.append(SearchResult(clean_job, keywords))
+                    filtered_jobs.append(SearchResult(job, []))
+            elif self.extraction_service.detect_keywords(job.description):
+                keywords = self.extraction_service.extract_keywords(job.description)
+                filtered_jobs.append(SearchResult(job, keywords))
 
         return filtered_jobs
     
@@ -59,36 +58,3 @@ class JobService:
         except Exception as e:
             print(e)
             return False
-
-    def process_job(self, job: Tag) -> Job:
-        title = job.find("a", class_="job-card-container__link").get_text(strip=True, separator=", ")
-        business = job.find("div", class_="artdeco-entity-lockup__subtitle").get_text(strip=True, separator=", ")
-        url = job.find("a", class_="job-card-container__link")["href"]
-        img = job.find("img")
-        logo_url = img["src"] if img else None
-        
-        job_link = self.base_link + url
-        self.driver.get(job_link)
-        time.sleep(2)
-        
-        loaded_job_page = self.driver.page_source
-        soup = BeautifulSoup(loaded_job_page, 'html.parser')
-        job_details_tag = soup.find('div', id='job-details')
-        description = job_details_tag.get_text(separator='\n', strip=True)
-        
-        job_details = self.get_details(soup)
-
-        return Job(title,description,business,url,logo_url, job_details)
-    
-    def get_details(self, soup: BeautifulSoup) -> JobDetails:
-        primary_desc = soup.find('div', class_='job-details-jobs-unified-top-card__primary-description-container')
-        spans = primary_desc.find_all('span', class_='tvm__text tvm__text--low-emphasis')
-        location = date = applications = 'Not found'
-        if spans:
-            if len(spans) > 0:
-                location = spans[0].get_text(strip=True)
-            if len(spans) > 2:
-                date = spans[2].get_text(strip=True)
-            if len(spans) > 4:
-                applications = spans[4].get_text(strip=True)
-        return JobDetails(location,date,applications)
